+++ internal/controller/pengambilan_obat_controller.go
@@ -3,6 +3,7 @@ package controller
 import (
 	"github.com/gofiber/fiber/v3"
 	"log"
+	"math"
 	"prb_care_api/internal/constant"
 	"prb_care_api/internal/middleware"
 	"prb_care_api/internal/model"
@@ -48,7 +49,11 @@ func (c *PengambilanObatController) Get(ctx fiber.Ctx) error {
 		log.Println(err.Error())
 		return fiber.ErrBadRequest
 	}
-	request.ID = id
+	if id < math.MinInt32 || id > math.MaxInt32 {
+		log.Println("value out of range for int32")
+		return fiber.ErrBadRequest
+	}
+	request.ID = int32(id)
 	response, err := c.PengambilanObatService.Get(ctx.Context(), request)
 	if err != nil {
 		log.Println(err.Error())
@@ -82,7 +87,11 @@ func (c *PengambilanObatController) Update(ctx fiber.Ctx) error {
 		log.Println(err.Error())
 		return fiber.ErrBadRequest
 	}
-	request.ID = id
+	if id < math.MinInt32 || id > math.MaxInt32 {
+		log.Println("value out of range for int32")
+		return fiber.ErrBadRequest
+	}
+	request.ID = int32(id)
 	if err := ctx.Bind().JSON(request); err != nil {
 		log.Println(err.Error())
 		return fiber.ErrBadRequest
@@ -108,7 +117,11 @@ func (c *PengambilanObatController) Delete(ctx fiber.Ctx) error {
 		log.Println(err.Error())
 		return fiber.ErrBadRequest
 	}
-	request.ID = id
+	if id < math.MinInt32 || id > math.MaxInt32 {
+		log.Println("value out of range for int32")
+		return fiber.ErrBadRequest
+	}
+	request.ID = int32(id)
 	if err := c.PengambilanObatService.Delete(ctx.Context(), request); err != nil {
 		log.Println(err.Error())
 		return err
@@ -127,7 +140,11 @@ func (c *PengambilanObatController) Batal(ctx fiber.Ctx) error {
 		log.Println(err.Error())
 		return fiber.ErrBadRequest
 	}
-	request.ID = id
+	if id < math.MinInt32 || id > math.MaxInt32 {
+		log.Println("value out of range for int32")
+		return fiber.ErrBadRequest
+	}
+	request.ID = int32(id)
 	if err := c.PengambilanObatService.Batal(ctx.Context(), request); err != nil {
 		log.Println(err.Error())
 		return err
@@ -146,7 +163,11 @@ func (c *PengambilanObatController) Diambil(ctx fiber.Ctx) error {
 		log.Println(err.Error())
 		return fiber.ErrBadRequest
 	}
-	request.ID = id
+	if id < math.MinInt32 || id > math.MaxInt32 {
+		log.Println("value out of range for int32")
+		return fiber.ErrBadRequest
+	}
+	request.ID = int32(id)
 	if err := c.PengambilanObatService.Diambil(ctx.Context(), request); err != nil {
 		log.Println(err.Error())
 		return err
